+++ src/layouts/mainLayout/MainLayout.tsx
@@ -1,11 +1,18 @@
 import { Outlet } from 'react-router';
+import { Suspense } from 'react';
 
 import NavigationMain from 'components/navigation/navigationMain';
+import Box from 'components/box';
+import Loader from 'components/loader';
 
 const MainLayout = () => (
   <>
     <NavigationMain />
-    <Outlet />
+    <Box as="main" bg="mainBackgroundColor">
+      <Suspense fallback={<Loader />}>
+        <Outlet />
+      </Suspense>
+    </Box>
   </>
 );
 export default MainLayout;
